#2696 - Document-level recommendations
- Removed a bunch of unused methods
- Moved a bunch of methods only used once to the classes where they are used
