+++ inception/inception-imls-opennlp/src/main/java/de/tudarmstadt/ukp/inception/recommendation/imls/opennlp/ner/OpenNlpNerRecommender.java
@@ -104,10 +104,10 @@ public class OpenNlpNerRecommender
         // OpenNLP
         int beamSize = Math.max(maxRecommendations, NameFinderME.DEFAULT_BEAM_SIZE);
 
-        TrainingParameters params = traits.getParameters();
+        var params = traits.getParameters();
         params.put(BeamSearch.BEAM_SIZE_PARAMETER, Integer.toString(beamSize));
 
-        TokenNameFinderModel model = train(nameSamples, params);
+        var model = train(nameSamples, params);
 
         aContext.put(KEY_MODEL, model);
     }
@@ -304,7 +304,7 @@ public class OpenNlpNerRecommender
         return label;
     }
 
-    private List<NameSample> extractNameSamples(List<CAS> aCasses)
+    private List<NameSample> extractNameSamples(Iterable<CAS> aCasses)
     {
         var nameSamples = new ArrayList<NameSample>();
 
@@ -402,8 +402,8 @@ public class OpenNlpNerRecommender
             TrainingParameters aParameters)
         throws RecommendationException
     {
-        try (NameSampleStream stream = new NameSampleStream(aNameSamples)) {
-            TokenNameFinderFactory finderFactory = new TokenNameFinderFactory();
+        try (var stream = new NameSampleStream(aNameSamples)) {
+            var finderFactory = new TokenNameFinderFactory();
             return NameFinderME.train("unknown", null, stream, aParameters, finderFactory);
         }
         catch (IOException e) {
