+++ inception/inception-pdf-editor/src/main/java/de/tudarmstadt/ukp/inception/pdfeditor/PdfAnnotationEditor.java
@@ -20,10 +20,10 @@ package de.tudarmstadt.ukp.inception.pdfeditor;
 import static de.tudarmstadt.ukp.inception.pdfeditor.pdfanno.render.PdfAnnoSerializer.convertToDocumentOffset;
 import static de.tudarmstadt.ukp.inception.pdfeditor.pdfanno.render.PdfAnnoSerializer.convertToDocumentOffsets;
 import static de.tudarmstadt.ukp.inception.rendering.vmodel.VID.NONE_ID;
-import static de.tudarmstadt.ukp.inception.support.uima.WebAnnoCasUtil.selectSentenceAt;
 import static de.tudarmstadt.ukp.inception.support.wicket.ServletContextUtils.referenceToUrl;
 import static java.lang.String.join;
 import static java.util.Arrays.asList;
+import static org.apache.uima.fit.util.CasUtil.getType;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -37,6 +37,7 @@ import javax.servlet.ServletContext;
 import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.apache.uima.cas.CAS;
 import org.apache.uima.cas.text.AnnotationFS;
+import org.apache.uima.fit.util.CasUtil;
 import org.apache.wicket.Component;
 import org.apache.wicket.ajax.AjaxRequestTarget;
 import org.apache.wicket.feedback.IFeedback;
@@ -48,6 +49,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.casstorage.CasProvider;
+import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Sentence;
 import de.tudarmstadt.ukp.inception.documents.api.DocumentService;
 import de.tudarmstadt.ukp.inception.editor.AnnotationEditorExtensionRegistry;
 import de.tudarmstadt.ukp.inception.editor.action.AnnotationActionHandler;
@@ -313,5 +315,21 @@ public class PdfAnnotationEditor
                 "'primary': true,", //
                 "'colorMap': {},", //
                 "'annotations':[annoFile]}, true);");
+    }
+
+    /**
+     * Get the sentence for this CAS based on the begin and end offsets. This is basically used to
+     * transform sentence address in one CAS to other sentence address for different CAS
+     *
+     * @param aCas
+     *            the CAS.
+     * @param aBegin
+     *            the begin offset.
+     * @return the sentence.
+     */
+    private static AnnotationFS selectSentenceAt(CAS aCas, int aBegin)
+    {
+        return CasUtil.select(aCas, getType(aCas, Sentence.class)).stream()
+                .filter(s -> s.getBegin() == aBegin).findFirst().orElse(null);
     }
 }
